Move simple string view out to common/string_view.h
